+++ toolium/driver_wrapper.py
@@ -16,64 +16,38 @@ See the License for the specific language governing permissions and
 limitations under the License.
 """
 
-import datetime
 import logging.config
 import os
 
 from toolium.config_driver import ConfigDriver
+from toolium.config_files import ConfigFiles
 from toolium.config_parser import ExtendedConfigParser
+from toolium.driver_wrappers_pool import DriverWrappersPool
+from toolium.utils import Utils
 
 
 class DriverWrapper(object):
-    # Singleton instance
-    _instance = None
-    is_additional = False
     driver = None
+    utils = None
+    session_id = None
+    remote_video_node = None
     logger = None
     config = ExtendedConfigParser()
-    browser_info = None
-    baseline_name = None
+
     # Configuration and output files
-    config_directory = None
-    output_directory = None
     config_properties_filenames = None
     config_log_filename = None
-    # Screenshots configuration
-    screenshots_directory = None
-    screenshots_number = None
-    # Videos configuration
-    videos_directory = None
-    videos_number = None
-    # Visual Testing configuration
-    visual_output_directory = None
     visual_baseline_directory = None
-    visual_number = None
+    baseline_name = None
 
-    def __new__(cls, *args, **kwargs):
-        if 'main_driver' in kwargs and kwargs['main_driver']:
-            if cls._instance:
-                # Return the singleton instance
-                instance = cls._instance
-            else:
-                # Create new instance and save it in a class property
-                instance = super(DriverWrapper, cls).__new__(cls)
-                cls._instance = instance
-        else:
-            # Create new instance with a copy of the config object
-            instance = super(DriverWrapper, cls).__new__(cls)
-            instance.config = cls._instance.config.deepcopy()
-            instance.is_additional = True
-        return instance
+    def __init__(self, main_driver=False):
+        if not main_driver:
+            # Copy config object from default driver
+            self.config = DriverWrappersPool.get_default_wrapper().config.deepcopy()
 
-    def get_configured_value(self, system_property_name, specific_value, default_value):
-        """Get configured value from system properties, method parameters or default value
-
-        :returns: configured value
-        """
-        try:
-            return os.environ[system_property_name]
-        except KeyError:
-            return specific_value if specific_value else default_value
+        # Create utils instance and add wrapper to the pool
+        self.utils = Utils(self)
+        DriverWrappersPool.add_wrapper(self)
 
     def configure_logger(self, tc_config_log_filename=None, tc_output_log_filename=None):
         """Configure selenium instance logger
@@ -82,10 +56,12 @@ class DriverWrapper(object):
         :param tc_output_log_filename: test case specific output logger file
         """
         # Get config and output logger files
-        config_log_filename = self.get_configured_value('Config_log_filename', tc_config_log_filename, 'logging.conf')
-        config_log_filename = os.path.join(self.config_directory, config_log_filename)
-        output_log_filename = self.get_configured_value('Output_log_filename', tc_output_log_filename, 'toolium.log')
-        output_log_filename = os.path.join(self.output_directory, output_log_filename)
+        config_log_filename = DriverWrappersPool.get_configured_value('Config_log_filename', tc_config_log_filename,
+                                                                      'logging.conf')
+        config_log_filename = os.path.join(DriverWrappersPool.config_directory, config_log_filename)
+        output_log_filename = DriverWrappersPool.get_configured_value('Output_log_filename', tc_output_log_filename,
+                                                                      'toolium.log')
+        output_log_filename = os.path.join(DriverWrappersPool.output_directory, output_log_filename)
         output_log_filename = output_log_filename.replace('\\', '\\\\')
 
         # Configure logger if logging filename has changed
@@ -102,8 +78,10 @@ class DriverWrapper(object):
 
         :param tc_config_prop_filenames: test case specific properties filenames
         """
-        prop_filenames = self.get_configured_value('Config_prop_filenames', tc_config_prop_filenames, 'properties.cfg')
-        prop_filenames = [os.path.join(self.config_directory, filename) for filename in prop_filenames.split(';')]
+        prop_filenames = DriverWrappersPool.get_configured_value('Config_prop_filenames', tc_config_prop_filenames,
+                                                                 'properties.cfg')
+        prop_filenames = [os.path.join(DriverWrappersPool.config_directory, filename) for filename in
+                          prop_filenames.split(';')]
         prop_filenames = ';'.join(prop_filenames)
 
         # Configure config if properties filename has changed
@@ -115,25 +93,8 @@ class DriverWrapper(object):
             # Override properties with system properties
             self.config.update_from_system_properties()
 
-    def configure_visual_directories(self):
-        """Configure screenshots, videos and visual directories"""
-        browser_info = self.config.get('Browser', 'browser').replace('-', '_')
-
-        # Configure directories if browser has changed
-        if self.browser_info != browser_info:
-            self.browser_info = browser_info
-
-            # Unique screenshots and videos directories
-            date = datetime.datetime.now().strftime('%Y-%m-%d_%H%M%S')
-            self.screenshots_directory = os.path.join(self.output_directory, 'screenshots', date + '_' + browser_info)
-            self.screenshots_number = 1
-            self.videos_directory = os.path.join(self.output_directory, 'videos', date + '_' + browser_info)
-            self.videos_number = 1
-
-            # Unique visualtests directories
-            self.visual_output_directory = os.path.join(self.output_directory, 'visualtests', date + '_' + browser_info)
-            self.visual_number = 1
-
+    def configure_visual_baseline(self):
+        """Configure baseline directory"""
         # Get baseline name
         baseline_name = self.config.get_optional('VisualTests', 'baseline_name', '{Browser_browser}')
         for section in self.config.sections():
@@ -144,46 +105,29 @@ class DriverWrapper(object):
         # Configure baseline directory if baseline name has changed
         if self.baseline_name != baseline_name:
             self.baseline_name = baseline_name
-            self.visual_baseline_directory = os.path.join(self.output_directory, 'visualtests', 'baseline',
-                                                          baseline_name)
+            self.visual_baseline_directory = os.path.join(DriverWrappersPool.output_directory, 'visualtests',
+                                                          'baseline', baseline_name)
 
-    def configure(self, is_selenium_test=True, tc_config_directory=None, tc_output_directory=None,
-                  tc_config_prop_filenames=None, tc_config_log_filename=None, tc_output_log_filename=None):
+    def configure(self, is_selenium_test=True, tc_config_files=ConfigFiles()):
         """Configure initial selenium instance using logging and properties files for Selenium or Appium tests
 
         :param is_selenium_test: true if test is a selenium or appium test case
-        :param tc_config_directory: test case specific config directory
-        :param tc_output_directory: test case specific output directory
-        :param tc_config_prop_filenames: test case specific properties filenames
-        :param tc_config_log_filename: test case specific logging config filename
-        :param tc_output_log_filename: test case specific output logger filename
+        :param tc_config_files: test case specific config files
         """
-        # Add warning message when old system properties are configured
-        deprecated = (('Files_properties', 'Config_directory and Config_prop_filenames'),
-                      ('Files_logging', 'Config_directory and Config_log_filename'),
-                      ('Files_log_filename', 'Output_directory and Output_log_filename'),
-                      ('Files_output_path', 'Output_directory'))
-        for prop in deprecated:
-            if self.get_configured_value(prop[0], None, None):
-                print('[WARN] {} system property is deprecated, use {} instead'.format(prop[0], prop[1]))
-
-        # Get config and output directories
-        self.config_directory = self.get_configured_value('Config_directory', tc_config_directory,
-                                                          os.path.join(os.getcwd(), 'conf'))
-        self.output_directory = self.get_configured_value('Output_directory', tc_output_directory,
-                                                          os.path.join(os.getcwd(), 'output'))
-        if not os.path.exists(self.output_directory):
-            os.makedirs(self.output_directory)
+        # Configure config and output directories
+        DriverWrappersPool.configure_common_directories(tc_config_files)
 
         # Configure logger
-        self.configure_logger(tc_config_log_filename, tc_output_log_filename)
+        self.configure_logger(tc_config_files.config_log_filename, tc_config_files.output_log_filename)
 
         # Initialize the config object
-        self.configure_properties(tc_config_prop_filenames)
+        self.configure_properties(tc_config_files.config_properties_filenames)
 
         # Configure visual directories
         if is_selenium_test:
-            self.configure_visual_directories()
+            browser_info = self.config.get('Browser', 'browser').replace('-', '_')
+            DriverWrappersPool.configure_visual_directories(browser_info)
+            self.configure_visual_baseline()
 
     def connect(self, maximize=True):
         """Set up the selenium driver and connect to the server
@@ -192,12 +136,13 @@ class DriverWrapper(object):
         :returns: selenium driver
         """
         self.driver = ConfigDriver(self.config).create_driver()
-        if self.is_additional:
-            from toolium.test_cases import SeleniumTestCase
-            SeleniumTestCase.additional_drivers.append(self.driver)
 
+        # Save session id and remote node to download video after the test execution
+        self.session_id = self.driver.session_id
+        self.remote_video_node = self.utils.get_remote_video_node()
+
+        # Maximize browser
         if maximize and self.is_maximizable():
-            # Maximize browser
             self.driver.maximize_window()
 
         return self.driver