+++ toolium/test/pageelements/test_page_elements.py
@@ -21,6 +21,8 @@ import unittest
 import mock
 from selenium.webdriver.common.by import By
 
+from toolium.driver_wrapper import DriverWrapper
+from toolium.driver_wrappers_pool import DriverWrappersPool
 from toolium.pageelements import *
 from toolium.pageelements import select_page_element
 from toolium.pageobjects.page_object import PageObject
@@ -66,8 +68,15 @@ class TestPageElements(unittest.TestCase):
         global mock_element
         mock_element = get_mock_element()
 
-    @mock.patch('toolium.toolium_wrapper.driver')
-    def test_locator(self, driver):
+        # Reset wrappers pool values
+        DriverWrappersPool._empty_pool()
+        DriverWrapper.config_properties_filenames = None
+
+        # Create a new wrapper
+        self.driver_wrapper = DriverWrappersPool.get_default_wrapper()
+        self.driver_wrapper.driver = mock.MagicMock()
+
+    def test_locator(self):
         page_object = LoginPageObject()
 
         self.assertEqual(page_object.title.locator, (By.ID, 'title'))
@@ -76,44 +85,37 @@ class TestPageElements(unittest.TestCase):
         self.assertEqual(page_object.language.locator, (By.ID, 'language'))
         self.assertEqual(page_object.login.locator, (By.ID, 'login'))
 
-    @mock.patch('toolium.toolium_wrapper.driver')
-    def test_get_text(self, driver):
-        driver.find_element.return_value = mock_element
+    def test_get_text(self):
+        self.driver_wrapper.driver.find_element.return_value = mock_element
 
         title_value = LoginPageObject().title.text
 
         self.assertEqual(title_value, 'text value')
 
-    @mock.patch('toolium.toolium_wrapper.driver')
-    def test_get_inputtext(self, driver):
-        driver.find_element.return_value = mock_element
+    def test_get_inputtext(self):
+        self.driver_wrapper.driver.find_element.return_value = mock_element
 
         username_value = LoginPageObject().username.text
 
         self.assertEqual(username_value, 'input text value')
 
-    @mock.patch('toolium.toolium_wrapper.driver')
-    def test_set_inputtext(self, driver):
-        # Configure driver ios mock
-        driver.find_element.return_value = mock_element
-        driver_ios_patch = mock.patch('toolium.toolium_wrapper.is_ios_test', mock.MagicMock(return_value=False))
-        driver_ios_patch.start()
+    def test_set_inputtext(self):
+        # Configure driver mock
+        self.driver_wrapper.driver.find_element.return_value = mock_element
+        self.driver_wrapper.is_ios_test = mock.MagicMock(return_value=False)
 
         LoginPageObject().username.text = 'new input value'
 
         self.assertEqual(mock_element.send_keys.mock_calls, [mock.call('new input value')])
-        driver_ios_patch.stop()
 
-    @mock.patch('toolium.toolium_wrapper.driver')
-    def test_get_selected_option(self, driver):
+    def test_get_selected_option(self):
         select_page_element.SeleniumSelect = get_mock_select()
 
         option = LoginPageObject().language.option
 
         self.assertEqual(option, 'option value')
 
-    @mock.patch('toolium.toolium_wrapper.driver')
-    def test_select_option(self, driver):
+    def test_select_option(self):
         select_page_element.SeleniumSelect = get_mock_select()
 
         LoginPageObject().language.option = 'new option value'
@@ -122,9 +124,8 @@ class TestPageElements(unittest.TestCase):
         self.assertEqual(select_page_element.SeleniumSelect().mock_calls,
                          [mock.call.select_by_visible_text('new option value')])
 
-    @mock.patch('toolium.toolium_wrapper.driver')
-    def test_click_button(self, driver):
-        driver.find_element.return_value = mock_element
+    def test_click_button(self):
+        self.driver_wrapper.driver.find_element.return_value = mock_element
         LoginPageObject().login.click()
 
         self.assertEqual(mock_element.click.mock_calls, [mock.call()])